Fixed big brother edit screen and layout
